+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.2 [2018-07-31]
+
+- Add `cacheFile` config option.
+
 ## 0.1.1 [2018-07-23]
 
 - Add `cache` config option.